modify customer

+++ src/main/java/com/se1889_jv/swp391/swpstart/service/PaymentService.java
@@ -62,7 +62,7 @@ public class PaymentService {
                         // If not, save it in DB with "PENDING" status
                         TransactionPayment transactionPayment = new TransactionPayment();
                         transactionPayment.setTransactionId(transactionId);
-                        transactionPayment.setAmount((Double) payment.get("amount"));
+                        transactionPayment.setAmount((Integer) payment.get("amount"));
                         transactionPayment.setContent((String) payment.get("content"));
 
                         String dateString = (String) payment.get("date");
